refactor: 일급 객체 내부에서 Ticket 추가하도록 변경
