+++ src/test/java/step2/LottoServiceTest.java
@@ -2,7 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step2.domain.LottoService;
+import step2.serivce.LottoService;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -18,6 +18,6 @@ public class LottoServiceTest {
     @DisplayName("로또 생성")
     @Test
     void generateAutoLottoTest() {
-        assertThat(lottoService.generateAutoLotto(3).getLottoTickets()).hasSize(3);
+        assertThat(lottoService.getLottoTickets(3).getLottoTickets()).hasSize(3);
     }
 }
